CustomerChurnAnalysis tamamlandı, manager, controller ve view kodları yazıldı. RegionalDemandForecast entitysi oluşturuldu ve dbcontext kayıdı tamamlandı. OnModelCreating kodları yazıldı.

+++ Core Layer/Dtos/ChurnPredictionDtos/ChurnPredictionResultDto.cs
@@ -14,7 +14,7 @@ namespace Core_Layer.Dtos.ChurnPredictionDtos
         public double ChurnProbability { get; set; }
         public float Recency { get; set; }
         public float Frequency { get; set; }
-        public float Monetary { get; set; }
+        public decimal Monetary { get; set; }
         public DateTime LastUpdated { get; set; }
         public string RiskLevel => ChurnProbability switch
         {
